+++ configserver/src/main/java/com/yahoo/vespa/config/server/SuperModelController.java
@@ -1,167 +1,86 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.vespa.config.server;
 
-import com.google.inject.Inject;
-import com.yahoo.cloud.config.ConfigserverConfig;
 import com.yahoo.config.ConfigInstance;
+import com.yahoo.config.ConfigurationRuntimeException;
+import com.yahoo.config.codegen.DefParser;
+import com.yahoo.config.codegen.InnerCNode;
 import com.yahoo.config.model.api.ConfigDefinitionRepo;
-import com.yahoo.config.provision.Version;
-import com.yahoo.config.provision.Zone;
-import com.yahoo.log.LogLevel;
-import com.yahoo.vespa.config.ConfigKey;
-import com.yahoo.vespa.config.GetConfigRequest;
-import com.yahoo.vespa.config.protocol.ConfigResponse;
-import com.yahoo.vespa.config.server.application.Application;
 import com.yahoo.config.provision.ApplicationId;
-import com.yahoo.config.provision.TenantName;
-import com.yahoo.vespa.config.GenerationCounter;
-import com.yahoo.vespa.config.server.application.ApplicationSet;
+import com.yahoo.vespa.config.ConfigDefinitionKey;
+import com.yahoo.vespa.config.ConfigKey;
+import com.yahoo.vespa.config.ConfigPayload;
+import com.yahoo.vespa.config.GetConfigRequest;
+import com.yahoo.vespa.config.buildergen.ConfigDefinition;
+import com.yahoo.vespa.config.protocol.ConfigResponse;
+import com.yahoo.vespa.config.protocol.DefContent;
 import com.yahoo.vespa.config.server.model.SuperModel;
 import com.yahoo.vespa.config.server.rpc.ConfigResponseFactory;
-import com.yahoo.vespa.config.server.rpc.ConfigResponseFactoryFactory;
 
 import java.io.IOException;
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
+import java.io.StringReader;
 
 /**
- * Controls the lifetime of the {@link SuperModel} and the {@link SuperModelRequestHandler}.
+ * Handler for global configs that must be resolved using the global SuperModel instance. Deals with
+ * reloading of config as well.
  *
  * @author lulf
  * @since 5.9
  */
-public class SuperModelController implements RequestHandler {
+public class SuperModelController {
 
-    private static final java.util.logging.Logger log = java.util.logging.Logger.getLogger(SuperModelController.class.getName());
-    private volatile SuperModelRequestHandler handler;
-    private final GenerationCounter generationCounter;
-    private final Zone zone;
-    private final long masterGeneration;
+    private final SuperModel model;
+    private final long generation;
     private final ConfigDefinitionRepo configDefinitionRepo;
     private final ConfigResponseFactory responseFactory;
-    private volatile boolean enabled = false;
 
-    /**
-     * Creates a supermodel controller
-     * 
-     * @param generationCounter this will be the SuperModelGenerationCounter in production
-     */
-    @Inject
-    public SuperModelController(GenerationCounter generationCounter, ConfigDefinitionRepo configDefinitionRepo, ConfigserverConfig configserverConfig) {
-        this.generationCounter = generationCounter;
+    public SuperModelController(SuperModel model, ConfigDefinitionRepo configDefinitionRepo, long generation, ConfigResponseFactory responseFactory) {
+        this.model = model;
         this.configDefinitionRepo = configDefinitionRepo;
-        this.masterGeneration = configserverConfig.masterGeneration();
-        this.responseFactory = ConfigResponseFactoryFactory.createFactory(configserverConfig);
-        this.zone = new Zone(configserverConfig);
-        this.handler = createNewHandler(Collections.emptyMap());
+        this.generation = generation;
+        this.responseFactory = responseFactory;
     }
 
     /**
-     * Signals that config has been reloaded for an {@link com.yahoo.vespa.config.server.application.Application}
-     * belonging to a tenant.
+     * Resolves global config for given request.
      *
-     * TODO: This is a bit too complex I think.
-     *
-     * @param tenant Name of tenant owning the application.
-     * @param applicationSet The reloaded set of {@link com.yahoo.vespa.config.server.application.Application}.
+     * @param request The {@link com.yahoo.vespa.config.GetConfigRequest} to find config for.
+     * @return a {@link com.yahoo.vespa.config.protocol.ConfigResponse} containing the response for this request.
+     * @throws java.lang.IllegalArgumentException if no such config was found.
      */
-    public synchronized void reloadConfig(TenantName tenant, ApplicationSet applicationSet) {
-        Map<TenantName, Map<ApplicationId, Application>> newModels = createModelCopy();
-        if (!newModels.containsKey(tenant)) {
-            newModels.put(tenant, new LinkedHashMap<>());
+    public ConfigResponse resolveConfig(GetConfigRequest request) {
+        ConfigKey<?> configKey = request.getConfigKey();
+        InnerCNode targetDef = getConfigDefinition(request.getConfigKey(), request.getDefContent());
+        try {
+            ConfigPayload payload = model.getConfig(configKey);
+            return responseFactory.createResponse(payload, targetDef, generation);
+        } catch (IOException e) {
+            throw new ConfigurationRuntimeException("Unable to resolve config", e);
         }
-        // TODO: Should supermodel care about multiple versions?
-        newModels.get(tenant).put(applicationSet.getId(), applicationSet.getForVersionOrLatest(Optional.empty()));
-        handler = createNewHandler(newModels);
     }
 
-    public synchronized void removeApplication(ApplicationId applicationId) {
-        Map<TenantName, Map<ApplicationId, Application>> newModels = createModelCopy();
-        if (newModels.containsKey(applicationId.tenant())) {
-            newModels.get(applicationId.tenant()).remove(applicationId);
-            if (newModels.get(applicationId.tenant()).isEmpty()) {
-                newModels.remove(applicationId.tenant());
+    private InnerCNode getConfigDefinition(ConfigKey<?> configKey, DefContent defContent) {
+        if (defContent.isEmpty()) {
+            ConfigDefinitionKey configDefinitionKey = new ConfigDefinitionKey(configKey.getName(), configKey.getNamespace());
+            ConfigDefinition configDefinition = configDefinitionRepo.getConfigDefinitions().get(configDefinitionKey);
+            if (configDefinition == null) {
+                throw new UnknownConfigDefinitionException("Unable to find config definition for '" + configKey.getNamespace() + "." + configKey.getName());
             }
-        }
-        handler = createNewHandler(newModels);
-    }
-
-    private SuperModelRequestHandler createNewHandler(Map<TenantName, Map<ApplicationId, Application>> newModels) {
-        long generation = generationCounter.get() + masterGeneration;
-        SuperModel model = new SuperModel(newModels, zone);
-        return new SuperModelRequestHandler(model, configDefinitionRepo, generation, responseFactory);
-    }
-
-    private Map<TenantName, Map<ApplicationId, Application>> getCurrentModels() {
-        if (handler != null) {
-            return handler.getSuperModel().getCurrentModels();
+            return configDefinition.getCNode();
         } else {
-            return new LinkedHashMap<>();
+            DefParser dParser = new DefParser(configKey.getName(), new StringReader(defContent.asString()));
+            return dParser.getTree();
         }
     }
 
-    private Map<TenantName, Map<ApplicationId, Application>> createModelCopy() {
-        Map<TenantName, Map<ApplicationId, Application>> currentModels = getCurrentModels();
-        Map<TenantName, Map<ApplicationId, Application>> newModels = new LinkedHashMap<>();
-        for (Map.Entry<TenantName, Map<ApplicationId, Application>> entry : currentModels.entrySet()) {
-            Map<ApplicationId, Application> appMap = new LinkedHashMap<>();
-            newModels.put(entry.getKey(), appMap);
-            for (Map.Entry<ApplicationId, Application> appEntry : entry.getValue().entrySet()) {
-                appMap.put(appEntry.getKey(), appEntry.getValue());
-            }
-        }
-        return newModels;
+    SuperModel getSuperModel() {
+        return model;
     }
 
-    public SuperModelRequestHandler getHandler() { return handler; }
-
-    @Override
-    public ConfigResponse resolveConfig(ApplicationId appId, GetConfigRequest req, Optional<Version> vespaVersion) {
-        log.log(LogLevel.DEBUG, "SuperModelController resolving " + req + " for app id '" + appId + "'");
-        if (handler != null) {
-            return handler.resolveConfig(req);
-        }
-        return null;
-    }
+    long getGeneration() { return generation; }
 
     public <CONFIGTYPE extends ConfigInstance> CONFIGTYPE getConfig(Class<CONFIGTYPE> configClass, ApplicationId applicationId, String configId) throws IOException {
-        return handler.getConfig(configClass, applicationId, configId);
+        return model.getConfig(configClass, applicationId, configId);
     }
 
-    @Override
-    public Set<ConfigKey<?>> listConfigs(ApplicationId appId, Optional<Version> vespaVersion, boolean recursive) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Set<ConfigKey<?>> listNamedConfigs(ApplicationId appId, Optional<Version> vespaVersion, ConfigKey<?> key, boolean recursive) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Set<ConfigKey<?>> allConfigsProduced(ApplicationId appId, Optional<Version> vespaVersion) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Set<String> allConfigIds(ApplicationId appID, Optional<Version> vespaVersion) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public boolean hasApplication(ApplicationId appId, Optional<Version> vespaVersion) {
-        return enabled && appId.equals(ApplicationId.global());
-    }
-
-    @Override
-    public ApplicationId resolveApplicationId(String hostName) {
-        return ApplicationId.global();
-    }
-
-    public void enable() {
-        enabled = true;
-    }
 }
