+++ configserver/src/main/resources/configserver-app/services.xml
@@ -12,7 +12,7 @@
     <component id="com.yahoo.vespa.config.server.modelfactory.ModelFactoryRegistry" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.SuperModelGenerationCounter" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.session.SessionPreparer" bundle="configserver" />
-    <component id="com.yahoo.vespa.config.server.SuperModelController" bundle="configserver" />
+    <component id="com.yahoo.vespa.config.server.SuperModelRequestHandler" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.StaticConfigDefinitionRepo" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.provision.HostProvisionerProvider" bundle="configserver" />
     <component id="com.yahoo.vespa.curator.Curator" bundle="configserver" />
