+++ configserver/src/test/java/com/yahoo/vespa/config/server/SuperModelControllerTest.java
@@ -1,29 +1,37 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.vespa.config.server;
 
-import com.yahoo.cloud.config.ConfigserverConfig;
+import com.yahoo.cloud.config.LbServicesConfig;
 import com.yahoo.config.model.application.provider.FilesApplicationPackage;
-import com.yahoo.config.provision.Version;
+import com.yahoo.config.provision.*;
+import com.yahoo.jrt.Request;
+import com.yahoo.vespa.config.ConfigKey;
+import com.yahoo.cloud.config.LbServicesConfig.Tenants.Applications;
+import com.yahoo.vespa.config.protocol.CompressionType;
+import com.yahoo.vespa.config.protocol.DefContent;
+import com.yahoo.vespa.config.protocol.JRTClientConfigRequestV3;
+import com.yahoo.vespa.config.protocol.JRTServerConfigRequestV3;
+import com.yahoo.vespa.config.protocol.Trace;
 import com.yahoo.vespa.config.server.application.Application;
-import com.yahoo.config.provision.ApplicationId;
-import com.yahoo.config.provision.TenantName;
-import com.yahoo.vespa.config.server.application.ApplicationSet;
+import com.yahoo.vespa.config.server.model.SuperModel;
 import com.yahoo.vespa.config.server.monitoring.MetricUpdater;
-import com.yahoo.vespa.curator.mock.MockCurator;
+import com.yahoo.vespa.config.server.rpc.UncompressedConfigResponseFactory;
 import com.yahoo.vespa.model.VespaModel;
 
 import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.TemporaryFolder;
 import org.xml.sax.SAXException;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Collections;
+import java.util.LinkedHashMap;
+import java.util.Map;
 import java.util.Optional;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.*;
+import static org.hamcrest.core.Is.is;
+import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
 
 /**
  * @author lulf
@@ -31,100 +39,93 @@ import static org.junit.Assert.*;
  */
 public class SuperModelControllerTest {
 
-    private static final File testApp = new File("src/test/resources/deploy/app");
-    private SuperModelGenerationCounter counter;
-    private SuperModelController controller;
-
-    @Rule
-    public TemporaryFolder folder = new TemporaryFolder();
+    private SuperModelController handler;
 
     @Before
-    public void setup() throws IOException {
-        counter = new SuperModelGenerationCounter(new MockCurator());
-        controller = new SuperModelController(counter,
-                                              new TestConfigDefinitionRepo(),
-                                              new ConfigserverConfig(new ConfigserverConfig.Builder()));
+    public void setupHandler() throws IOException, SAXException {
+        Map<TenantName, Map<ApplicationId, Application>> models = new LinkedHashMap<>();
+        models.put(TenantName.from("a"), new LinkedHashMap<>());
+        File testApp = new File("src/test/resources/deploy/app");
+        ApplicationId app = ApplicationId.from(TenantName.from("a"),
+                                               ApplicationName.from("foo"), InstanceName.defaultName());
+        models.get(app.tenant()).put(app, new Application(new VespaModel(FilesApplicationPackage.fromFile(testApp)), new ServerCache(), 4l, Version.fromIntValues(1, 2, 3), MetricUpdater.createTestUpdater(), app));
+        handler = new SuperModelController(new SuperModel(models, Zone.defaultZone()), new TestConfigDefinitionRepo(), 2, new UncompressedConfigResponseFactory());
+    }
+    
+    @Test
+    public void test_lb_config_simple() {
+        LbServicesConfig.Builder lb = new LbServicesConfig.Builder();
+        handler.getSuperModel().getConfig(lb);
+        LbServicesConfig lbc = new LbServicesConfig(lb);
+        assertThat(lbc.tenants().size(), is(1));
+        assertThat(lbc.tenants("a").applications().size(), is(1));
+        Applications app = lbc.tenants("a").applications("foo:prod:default:default");
+        assertTrue(app.hosts().size() > 0);
+    }
+
+
+    @Test(expected = UnknownConfigDefinitionException.class)
+    public void test_unknown_config_definition() {
+        String md5 = "asdfasf";
+        Request request = JRTClientConfigRequestV3.createWithParams(new ConfigKey<>("foo", "id", "bar", md5, null), DefContent.fromList(Collections.emptyList()),
+                                                                    "fromHost", md5, 1, 1, Trace.createDummy(), CompressionType.UNCOMPRESSED,
+                                                                    Optional.empty())
+                                                  .getRequest();
+        JRTServerConfigRequestV3 v3Request = JRTServerConfigRequestV3.createFromRequest(request);
+        handler.resolveConfig(v3Request);
     }
 
     @Test
-    public void test_super_model_reload() throws IOException, SAXException {
-        TenantName tenantA = TenantName.from("a");
-        assertNotNull(controller.getHandler());
-        long gen = counter.increment();
-        controller.reloadConfig(tenantA, createApp(tenantA, "foo", 3l, 1));
-        assertNotNull(controller.getHandler());
-        assertThat(controller.getHandler().getGeneration(), is(gen));
-        controller.reloadConfig(tenantA, createApp(tenantA, "foo", 4l, 2));
-        assertThat(controller.getHandler().getGeneration(), is(gen));
-        // Test that a new app is used when there already exist an application with the same id
-        ApplicationId appId = new ApplicationId.Builder().tenant(tenantA).applicationName("foo").build();
-        assertThat(((TestApplication) controller.getHandler().getSuperModel().getCurrentModels().get(tenantA).get(appId)).version, is(2l));
-        gen = counter.increment();
-        controller.reloadConfig(tenantA, createApp(tenantA, "bar", 2l, 3));
-        assertThat(controller.getHandler().getGeneration(), is(gen));
+    public void test_lb_config_multiple_apps() throws IOException, SAXException {
+        Map<TenantName, Map<ApplicationId, Application>> models = new LinkedHashMap<>();
+        models.put(TenantName.from("t1"), new LinkedHashMap<>());
+        models.put(TenantName.from("t2"), new LinkedHashMap<>());
+        File testApp1 = new File("src/test/resources/deploy/app");
+        File testApp2 = new File("src/test/resources/deploy/advancedapp");
+        File testApp3 = new File("src/test/resources/deploy/advancedapp");
+        // TODO must fix equals, hashCode on Tenant
+        Version vespaVersion = Version.fromIntValues(1, 2, 3);
+        models.get(TenantName.from("t1")).put(applicationId("mysimpleapp"),
+                new Application(new VespaModel(FilesApplicationPackage.fromFile(testApp1)), new ServerCache(), 4l, vespaVersion, MetricUpdater.createTestUpdater(), applicationId("mysimpleapp")));
+        models.get(TenantName.from("t1")).put(applicationId("myadvancedapp"),
+                new Application(new VespaModel(FilesApplicationPackage.fromFile(testApp2)), new ServerCache(), 4l, vespaVersion, MetricUpdater.createTestUpdater(), applicationId("myadvancedapp")));
+        models.get(TenantName.from("t2")).put(applicationId("minetooadvancedapp"),
+                new Application(new VespaModel(FilesApplicationPackage.fromFile(testApp3)), new ServerCache(), 4l, vespaVersion, MetricUpdater.createTestUpdater(), applicationId("minetooadvancedapp")));
+
+        SuperModelController han = new SuperModelController(new SuperModel(models, Zone.defaultZone()), new TestConfigDefinitionRepo(), 2, new UncompressedConfigResponseFactory());
+        LbServicesConfig.Builder lb = new LbServicesConfig.Builder();
+        han.getSuperModel().getConfig(lb);
+        LbServicesConfig lbc = new LbServicesConfig(lb);
+        assertThat(lbc.tenants().size(), is(2));
+        assertThat(lbc.tenants("t1").applications().size(), is(2));
+        assertThat(lbc.tenants("t2").applications().size(), is(1));
+        assertThat(lbc.tenants("t2").applications("minetooadvancedapp:prod:default:default").hosts().size(), is(1));
+        assertQrServer(lbc.tenants("t2").applications("minetooadvancedapp:prod:default:default"));
     }
 
-    @Test
-    public void test_super_model_remove() throws IOException, SAXException {
-        TenantName tenantA = TenantName.from("a");
-        TenantName tenantB = TenantName.from("b");
-        long gen = counter.increment();
-        controller.reloadConfig(tenantA, createApp(tenantA, "foo", 3l, 1));
-        controller.reloadConfig(tenantA, createApp(tenantA, "bar", 30l, 2));
-        controller.reloadConfig(tenantB, createApp(tenantB, "baz", 9l, 3));
-        assertThat(controller.getHandler().getGeneration(), is(gen));
-        assertThat(controller.getHandler().getSuperModel().getCurrentModels().size(), is(2));
-        assertThat(controller.getHandler().getSuperModel().getCurrentModels().get(TenantName.from("a")).size(), is(2));
-        controller.removeApplication(
-                new ApplicationId.Builder().tenant("a").applicationName("unknown").build());
-        assertThat(controller.getHandler().getGeneration(), is(gen));
-        assertThat(controller.getHandler().getSuperModel().getCurrentModels().size(), is(2));
-        assertThat(controller.getHandler().getSuperModel().getCurrentModels().get(TenantName.from("a")).size(), is(2));
-        gen = counter.increment();
-        controller.removeApplication(
-                new ApplicationId.Builder().tenant("a").applicationName("bar").build());
-        assertThat(controller.getHandler().getSuperModel().getCurrentModels().size(), is(2));
-        assertThat(controller.getHandler().getSuperModel().getCurrentModels().get(TenantName.from("a")).size(), is(1));
-        assertThat(controller.getHandler().getGeneration(), is(gen));
+    private ApplicationId applicationId(String applicationName) {
+        return ApplicationId.from(TenantName.defaultName(),
+                                  ApplicationName.from(applicationName), InstanceName.defaultName());
     }
 
-    @Test
-    public void test_super_model_master_generation() throws IOException, SAXException {
-        TenantName tenantA = TenantName.from("a");
-        long masterGen = 10;
-        controller = new SuperModelController(counter,
-                                              new TestConfigDefinitionRepo(), 
-                                              new ConfigserverConfig(new ConfigserverConfig.Builder().masterGeneration(masterGen)));
-
-        long gen = counter.increment();
-        controller.reloadConfig(tenantA, createApp(tenantA, "foo", 3L, 1));
-        assertThat(controller.getHandler().getGeneration(), is(masterGen + gen));
-    }
-
-    @Test
-    public void test_super_model_has_application_when_enabled() {
-        assertFalse(controller.hasApplication(ApplicationId.global(), Optional.empty()));
-        controller.enable();
-        assertTrue(controller.hasApplication(ApplicationId.global(), Optional.empty()));
-    }
-
-    private ApplicationSet createApp(TenantName tenant, String application, long generation, long version) throws IOException, SAXException {
-        return ApplicationSet.fromSingle(
-                new TestApplication(
-                        new VespaModel(FilesApplicationPackage.fromFile(testApp)),
-                        new ServerCache(),
-                        generation,
-                        new ApplicationId.Builder().tenant(tenant).applicationName(application).build(),
-                        version));
-    }
-
-    private static class TestApplication extends Application {
-        private long version = 0;
-
-        public TestApplication(VespaModel vespaModel, ServerCache cache, long appGeneration, ApplicationId app, long version) {
-            super(vespaModel, cache, appGeneration, Version.fromIntValues(1, 2, 3), MetricUpdater.createTestUpdater(), app);
-            this.version = version;
+    private void assertQrServer(Applications app) {
+        String host = app.hosts().keySet().iterator().next();
+        Applications.Hosts hosts = app.hosts(host);
+        assertThat(hosts.hostname(), is(host));
+        for (Map.Entry<String, Applications.Hosts.Services> e : app.hosts(host).services().entrySet()) {
+            System.out.println(e);
+            if ("qrserver".equals(e.getKey())) {
+                Applications.Hosts.Services s = e.getValue();
+                assertThat(s.type(), is("qrserver"));
+                assertThat(s.ports().size(), is(4));
+                assertThat(s.index(), is(0));
+                return;
+            }
         }
+        org.junit.Assert.fail("No qrserver service in config");
     }
 
 }
+
+
+
